+++ cypress/integration/WidgetsForEnrollmentPages/WidgetsForEnrollmentEditEvent/index.js
@@ -1,6 +1,7 @@
 import '../sharedSteps';
 import '../WidgetEnrollment';
 import '../WidgetProfile';
+import '../WidgetEventComment';
 
 beforeEach(() => {
     cy.loginThroughForm();