+++ src/core_modules/capture-core-utils/rulesEngine/rulesEngine.types.js
@@ -175,9 +175,9 @@ export type TrackedEntityAttributes = {
 };
 
 export type Enrollment = {
-    enrollmentDate?: ?string,
-    incidentDate?: ?string,
-    enrollmentId?: ?string,
+    enrollmentDate?: string,
+    incidentDate?: string,
+    enrollmentId?: string,
 };
 
 export type TEIValues = {