+++ src/core_modules/capture-core/components/Pages/StageEvent/StageEventList/StageEventHeader/StageEventHeader.types.js
@@ -1,10 +1,9 @@
 // @flow
 import { Icon } from 'capture-core/metaData';
-import type { ApiTEIEvent } from 'capture-core/events/getEnrollmentEvents';
 
 export type Props = {|
   ...CssClasses,
   title: string,
-  events: Array<ApiTEIEvent>,
+  events: Array<ApiEnrollmentEvent>,
   icon?: Icon,
 |}