+++ src/core_modules/capture-core/components/WidgetEventComment/index.js
@@ -1,0 +1,1 @@
+export { WidgetEventComment } from './WidgetEventComment.component';