+++ src/core_modules/capture-core/components/WidgetEventEdit/DataEntry/editEventDataEntry.actions.js
@@ -23,8 +23,6 @@ export const batchActionTypes = {
     UPDATE_DATA_ENTRY_FIELD_EDIT_SINGLE_EVENT_ACTION_BATCH: 'UpdateDataEntryFieldForEditSingleEventActionsBatch',
     UPDATE_FIELD_EDIT_SINGLE_EVENT_ACTION_BATCH: 'UpdateFieldForEditSingleEventActionsBatch',
     RULES_EFFECTS_ACTIONS_BATCH: 'RulesEffectsForEditSingleEventActionsBatch',
-    ADD_NOTE_FOR_EDIT_SINGLE_EVENT_BATCH: 'AddNoteForEditSingleEventBatch',
-    REMOVE_NOTE_FOR_EDIT_SINGLE_EVENT_BATCH: 'RemoveNoteForEditSingleEventBatch',
 };
 
 export const actionTypes = {
@@ -38,10 +36,6 @@ export const actionTypes = {
     START_CANCEL_SAVE_RETURN_TO_MAIN_PAGE: 'CancelUpdateForSingleEventReturnToMainPage',
     NO_WORKING_LIST_UPDATE_NEEDED_AFTER_CANCEL_UPDATE: 'NoWorkingListUpdateNeededAfterEventUpdateCancelled',
     UPDATE_WORKING_LIST_AFTER_CANCEL_UPDATE: 'UpdateWorkingListAfterEventUpdateCancelled',
-    REQUEST_ADD_NOTE_FOR_EDIT_SINGLE_EVENT: 'RequestAddNoteForEditSingleEvent',
-    START_ADD_NOTE_FOR_EDIT_SINGLE_EVENT: 'StartAddNoteForEditSingleEvent',
-    NOTE_ADDED_FOR_EDIT_SINGLE_EVENT: 'NoteAddedForEditSingleEvent',
-    ADD_NOTE_FAILED_FOR_EDIT_SINGLE_EVENT: 'AddNoteFailedForEditSingleEvent',
 };
 
 export const editEventIds = {
@@ -126,7 +120,7 @@ export const openEventForEditInDataEntry = (
     const stage = program instanceof TrackerProgram ? getStageFromEvent(eventContainer.event)?.stage : undefined;
     const allEventsData = program instanceof EventProgram || !allEvents
         ? [eventDataForRulesEngine]
-        : [...prepareEnrollmentEventsForRulesEngine(eventDataForRulesEngine, allEvents)];
+        : [...prepareEnrollmentEventsForRulesEngine(allEvents, eventDataForRulesEngine)];
 
     return [
         ...dataEntryActions,
@@ -181,18 +175,3 @@ export const startAsyncUpdateFieldForEditEvent = (
 ) =>
     actionPayloadAppender(innerAction)({ onSuccess, onError });
 
-export const requestAddNoteForEditSingleEvent = (itemId: string, dataEntryId: string, note: string) =>
-    actionCreator(actionTypes.REQUEST_ADD_NOTE_FOR_EDIT_SINGLE_EVENT)({ itemId, dataEntryId, note });
-
-export const startAddNoteForEditSingleEvent = (eventId: string, serverData: Object, selections: Object, context: Object) =>
-    actionCreator(actionTypes.START_ADD_NOTE_FOR_EDIT_SINGLE_EVENT)({ selections, context }, {
-        offline: {
-            effect: {
-                url: `events/${eventId}/note`,
-                method: effectMethods.POST,
-                data: serverData,
-            },
-            commit: { type: actionTypes.NOTE_ADDED_FOR_EDIT_SINGLE_EVENT, meta: { selections, context } },
-            rollback: { type: actionTypes.ADD_NOTE_FAILED_FOR_EDIT_SINGLE_EVENT, meta: { selections, context } },
-        },
-    });