+++ src/core_modules/capture-core/components/WidgetEventEdit/DataEntry/epics/editEventDataEntry.epics.js
@@ -44,7 +44,7 @@ export const openEditEventInDataEntryEpic = (action$: InputObservable, store: Re
             const program = metadataContainer.program;
 
 
-            return batchActions(openEventForEditInDataEntry(eventContainer, orgUnit, foundation, program, state.enrollmentSite?.events));
+            return batchActions(openEventForEditInDataEntry(eventContainer, orgUnit, foundation, program, state.enrollmentDomain.enrollment?.events));
         }));
 
 
@@ -68,9 +68,9 @@ const runRulesForEditSingleEvent = (store: ReduxStore, dataEntryId: string, item
     let currentEventMainData = foundation ? getCurrentClientMainData(state, itemId, dataEntryId, foundation) : {};
     currentEventMainData = { ...state.events[eventId], ...currentEventMainData };
     const currentEventData = { ...currentEventValues, ...currentEventMainData };
-    const allEvents = state.enrollmentSite?.events;
+    const allEvents = state.enrollmentDomain.enrollments?.events;
     const allEventsData = program instanceof TrackerProgram && allEvents
-        ? [...prepareEnrollmentEventsForRulesEngine(currentEventData, allEvents)]
+        ? [...prepareEnrollmentEventsForRulesEngine(allEvents, currentEventData)]
         : [currentEventData];
 
     const rulesActions = getRulesActionsForEvent(