+++ src/core_modules/capture-core/components/WidgetEventEdit/EditEventDataEntry/EditEventDataEntry.component.js
@@ -261,7 +261,6 @@ type Props = {
     onStartAsyncUpdateField: Object,
     onSave: (eventId: string, dataEntryId: string, formFoundation: RenderFoundation) => void,
     onCancel: () => void,
-    onAddNote: (itemId: string, dataEntryId: string, note: string) => void,
     classes: {
         dataEntryContainer: string,
         fieldLabelMediaBased?: ?string,