+++ src/core_modules/capture-core/components/WidgetEventEdit/ViewEventDataEntry/viewEventDataEntry.epics.js
@@ -61,7 +61,7 @@ export const loadViewEventDataEntryEpic: Epic = (action$, store) =>
             }
             const foundation = metadataContainer.stage.stageForm;
             const program = metadataContainer.program;
-            return from(loadViewEventDataEntry(eventContainer, orgUnit, foundation, program, state.enrollmentSite?.events))
+            return from(loadViewEventDataEntry(eventContainer, orgUnit, foundation, program, state.enrollmentDomain.enrollment?.events))
                 .pipe(
                     map(item => batchActions(item)),
                 );